Move files out of pages folder, move details/index.tsx to details.tsx
