+++ src/pageComponents/details/Chart.tsx
