+++ src/pageComponents/details/GitHubMetricIcon.tsx
