+++ src/pageComponents/details/ScatterPlot.tsx
