+++ src/pageComponents/details/Sidebar.tsx
